requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.7
httpx==0.25.1
apscheduler==3.10.4
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
kiteconnect==3.9.6
dhanhq==2.1.0
pytz==2023.3
requests==2.31.0
psutil==5.9.6
email-validator==2.1.0
